requirements.txt
amqp==5.3.1
billiard==4.2.1
blinker==1.9.0
celery==5.4.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
Flask==3.1.0
fonttools==4.55.8
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
kiwisolver==1.4.8
kombu==5.4.2
MarkupSafe==3.0.2
matplotlib==3.10.0
nltk==3.9.1
numpy==2.2.2
packaging==24.2
pandas==2.2.3
pillow==11.1.0
prompt_toolkit==3.0.50
pyparsing==3.2.1
python-dateutil==2.9.0.post0
pytz==2024.2
regex==2024.11.6
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
six==1.17.0
textblob==0.19.0
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2025.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
python-dotenv==1.0.0

